Make progress bar show empty on start of page load
Previously, when clicking on a link or opening a new URL, the user would
not get any feedback showing that the command was taken until after a
progress change, which could sometimes take a prolonged period of time.
Now, the user is immediately alerted via the statusbar going to zero.