Added *result to get data back from commands
This allows socket (and stdio) to have a way to reply back.
Had to change all their function definitions, unfortunately.

Currently, no functions write a result. print will, at least, probably
run_js too, and some others.

When a function calling parse_command (or parse_cmd_line) is not
interested in the result, it is safe to pass NULL. Before calling the
function, parse_command creates a GString and prints the result to
stdout if verbose is enabled. You could do the same when calling most
command functions directly, but not with those that expect a GString.

Oh, additionally, I killed NOSPLIT, moved the typedef Command to uzbl.h,
and created the struct CommandInfo which replaces a Command[2] array.
That fixes FS#22, which is a warning with -pedantic. Heh.