talk_to_socket: split arguments in packet with \0
In the packet that is written to the daemon via talk_to_socket(), we were
previously using a space to separate each argument, and asking uzbl to quote
the arguments as though for the shell. This just makes life complicated since
we have to undo this in the daemon.

This commit changes the behaviour so that arguments are delimited by ASCII
nul (\0) in the packet. The daemon can thus unambiguously determine where
each argument ends and the next begins.