Fixed `run_handler (const gchar *act, const gchar *args)` by checking if `act` is not `NULL` or empty.
It seems that there are a number of "hooks" that are executed when something happens (like accessing an url),
but usually by default these hooks are not set (thus NULL), and g_strsilit gives an NULL assertion warning.

Also we should check if `act` was not an empty string, thus `parts[0]` beeing `NULL`.