Consistency fix in the bind plugin with regard to '*' binds.
The problem (as highlighted by the link following scripts) was that when
the user enters a stack bind (like a "fl<link:>*" bind) that binds command
isn't executed with a null argument. When you do the equivalent without
stack binds (like a "fl*" bind) that binds command is executed with a
null argument as soon as you type "fl". This behaviour is useful and
exploited heavily by the link following scripts which use the null argument
as a trigger to "numberize" all the links in the visible area.