Made main() method for uzbl-event-manager
Calls main() upon running the event manager.  Work done was moving a
gigantic block of repetetive adding for the parser into its own function
`make_parser()`, moved `get_logger(...)` out of the if-main branch, and
declaring a lot of variables as globals (this can be fixed later;
recommended to make all globals in all caps later).

This allows someone to attach a debugger or profiler to the event
manager more easily, just by editing the script.