Ensures that UZBL_URI is set when window is mapped
This will help window manager to sort windows when they are mapped,
instead of moving them back and forth when they are already shown. This
is theoretically incurs tiny delay of showing window, but it's shown
before main loop is started anyway, and you can't see anything useful
before, so the delay is negligible.