introduce getter and setter functions.
some variables didn't always have a value reflecting the browser's
internal state. for example, if `default_encoding` was never set then
`print @default_encoding` would always print a blank string.

this introduces getter functions that ensure the value of a variable is
always in sync with the internal state of the browser.

also: setters, because sometimes you need to process user input
before storing it.