Configuration module refactored to new plugin system
Continuing one plugin at a time style. But this time
`PluginClass.get` conflicts with `get` inside config
plugin itself interface (which is in turn supports
dict-style methods). So after some discussion with @keis
and @bct on irc we came up with `PluginClass[uzbl]`
syntax as concise and clear way of finding plugin
instances.