Clean up the formfiller
This change breaks previous form filler files.

The formfiller needs some work. The following rework stores each profile
in its own file instead of all in one file. This allows the script to no
longer have to do grepping for a range of profile names and such.
Instead, a glob can handle this for us.