use asynchronous http auth
This gives more freedom in how to write auth handlers

Implemented as a event/command pair: AUTHENTICATE and auth. When a
request needs authentication AUTHENTICATE will be sent and the message
paused. The user then uses the auth command to authenticate and all
messages waiting for that authentication will be resumed.

* updated auth.py to send a auth command instead of answering on stdout
* updated example config to spawn auth.py using on_event AUTHENTICATE