Add a variable to mediate multiple button clicks
The keyevent is always sent for multiple clicks (that aren't in an edit
field for left and middle clicks), but the propogation of the event
through to webkit can be suppressed if the user wants something else to
happen.

This allows double and triple click selections (by word and paragraph,
respectively) work again.