Rewrite js command
Instead of having a "js" and "script" command, have a single js command
with a "context" parameter. Also included is an uzbl context, separate
from the webpage. This should allow scripts and such to be written in JS
once some API functions are written.

Also adds WebKit2 support (but it's broken in that there doesn't seem to
be a way to get the JS context where the DOM and such live).