Rewrite download support
The previous way of doing downloads required the callback function to be
exposed from gui.h. The new way fixes this because in WebKit2, a signal
gets sent for any download request. For WebKit1, a simpler API is
exposed in the gui.h header (without overloading the data pointer of the
callback).