Rewrite zoom_* commands
Instead of zoom_in and zoom_out commands, a single zoom command with
subcommands is implemented instead. The ability to set the zoom level to
a specific value is also added as well as one-shot zoom-step values.

Since WebKit2 does not have a zoom-step anywhere, it is simpler to do
the same for WebKit1 as well since the APIs are otherwise the same.