Toss out AST parsing for argument splitting
Using the AST to parse just makes things have to care that the EM is
written in Python3 and as such have extra escapes to worry about.

As an example, the following (as interpreted from the config file) fails
with the AST parser:

    @cbind  \\u  = blah

Because "\u  " is not a valid Unicode escape. Ideally, Python3 would
still have the string_escape codec to allow us to just do what we did in
Python2, but we're stuck with a manual regex replacement.