Update README
Remove the < and > characters from the format description of the geometry variable.  In the website version of the README the geometry format displays as "x++" instead of "WidthxHeight+xOffset+yOffset" because of the < and > characters, despite what the DaringFireball Markdown docs say.  Other markup to signify that Width etcetera are placeholders for numeric values will probably not render sensibly in every output format, so that understanding requires intelligence from the reader.