extio: initialize GError with `NULL`
Glib/GIO functions expect the error out parameters to be initialized to
`NULL`, but we fail to initialize it in `read_header_read_all_cb`. This
can actually cause undefined behavior and segfaults when executed, as is
observed for example when starting up `uzbl-core` with the uzbl
webextension.

Fix the error by initializing the variable to `NULL`.