gui: fix detection of click target
When performing a click inside of the WebView, we notify event handlers
whether a form or the root document is active. We do so by retrieving
inspecting the hit test result context, which includes a bit mask of
what has been hit by the click. The hit test result context is retrieved
via the function `get_click_context`. But upon further inspection, this
function will always return `NO_CLICK_CONTEXT`, breaking our ability to
distinguish between clicks on a form and the root document.

The WebKitHitTestResult can be acquired by listening on the
"mouse-target-changed" event. While we do so already, we do not store
the bitmask anywhere and as such, it is not available when
`get_click_context` is called.

We fix this by introducing a new field `last_hit_context` to the
`_UzblGui` structure, which is then subsequently used in
`get_click_context` to return the actual context. This fixes detection
of clicks on the document root.