Handle partial writes and EAGAIN failures in write_to_socket
g_io_channel_write_chars can not only succeed or fail completely, but
also do a partial write or return G_IO_STATUS_AGAIN when the
underlying socket / file descriptor is non-blocking and not ready to
send.  This is more likely to happen when a lot needs to be send, for
example long paths / URIs.